Create Practice1.py
Write a program which repeatedly reads numbers until the user enters "exit". Once "exit" is entered, print out the sum, count, and average of the numbers. If the user enters anything other than a number, detect their mistake using try and except and print an error message and skip to the next number.